Move Coaching Dashboard to its own tsx file
